requirements.txt
Scrapy==0.24
scrapyd
kombu==3.0.12
Pillow
lxml
cssselect
jieba
beautifulsoup
nltk
mongoengine
flask-restful>=0.2.12
requests
rq
redis
